+++ packages/core-js-compat/src/data.mjs
@@ -101,7 +101,7 @@ export const data = {
   'es.array.at': {
     chrome: '92',
     firefox: '90',
-    safari: '15.1',
+    safari: '15.4',
   },
   'es.array.concat': {
     chrome: '51',
@@ -726,7 +726,7 @@ export const data = {
   'es.object.has-own': {
     chrome: '93',
     firefox: '92',
-    safari: '15.1',
+    safari: '15.4',
   },
   'es.object.is': {
     android: '4.1',
@@ -983,7 +983,7 @@ export const data = {
   'es.string.at-alternative': {
     chrome: '92',
     firefox: '90',
-    safari: '15.1',
+    safari: '15.4',
   },
   'es.string.code-point-at': {
     chrome: '41',
@@ -1257,7 +1257,7 @@ export const data = {
   'es.typed-array.at': {
     chrome: '92',
     firefox: '90',
-    safari: '15.1',
+    safari: '15.4',
   },
   'es.typed-array.copy-within': {
     chrome: '45',
@@ -1442,10 +1442,10 @@ export const data = {
   'esnext.array.filter-reject': {
   },
   'esnext.array.find-last': {
-    safari: '15.1',
+    safari: '15.4',
   },
   'esnext.array.find-last-index': {
-    safari: '15.1',
+    safari: '15.4',
   },
   'esnext.array.group-by': {
   },
@@ -1718,10 +1718,10 @@ export const data = {
   'esnext.typed-array.filter-reject': {
   },
   'esnext.typed-array.find-last': {
-    safari: '15.1',
+    safari: '15.4',
   },
   'esnext.typed-array.find-last-index': {
-    safari: '15.1',
+    safari: '15.4',
   },
   'esnext.typed-array.group-by': {
   },